% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{Btot_cf}
\alias{Btot_cf}
\title{Total boron concentration in the calcifying fluid}
\usage{
Btot_cf(Sal)
}
\arguments{
\item{Sal}{The salinity in UPS}
}
\value{
The total boron concentration
}
\description{
Calculates the concentration of toal boron in the calcifying fluid in mmol/kg based on Zeebe and Wolf-Gladrow 2001
}
