% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{coralCF}
\alias{coralCF}
\title{Master function that calculated all parameters of Coral Calcifying Fluid and Seawater Chemistry}
\usage{
coralCF(
  TCel,
  Sal,
  d11B,
  BCa,
  d11Bsw = 39.5,
  alpha_klochko = 1.0272,
  species = c("Cladocora caespitosa", "Porites sp.", "Porites cylindrica",
    "Acropora nobilis", "Acropora sp.", "Stylophora pistillata"),
  ca_sw = 0.0112
)
}
\arguments{
\item{TCel}{Temperature in degrees Celsius.}

\item{Sal}{Salinity in UPS.}

\item{d11B}{The d11B of the coral skeleton (in permil).}

\item{BCa}{The B/Ca ratio of the coral skeleton (in mmol/mol).}

\item{d11Bsw}{Seawater d11B (in permil). Default is 39.5 (Foster 2008). Alternative is 39.61 (D'Olivo and McCulloch 2017).}

\item{alpha_klochko}{Isotopic fractionation factor alpha. Default is 1.0272 (Klochko et al. 2006).}

\item{species}{Coral species for pH reconstruction. Defaults to "Cladocora caespitosa".}

\item{ca_sw}{Concentration of Ca ions in seawater (mol/kg). Default is 0.0112 (McCulloch et al. 2012).}
}
\value{
A \code{data.frame} (or vector if single inputs are provided) containing all calculated 
  constants, pH values, ion concentrations, and saturation states.
}
\description{
Calculates the complete set of CF carbonate chemistry and seawater pH reconstruction from parameters from raw experimental and environmental data in a single step.
}
\examples{
# Calculate for single values
coralCF(TCel = 25, Sal = 35, d11B = 24.43, BCa = 664.8)

# Vectorized calculation for multiple samples
df_results <- coralCF(
  TCel = c(25, 26), 
  Sal = c(35, 36), 
  d11B = c(24.43, 25.10), 
  BCa = c(664.8, 670.2)
)
}
