% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{K1}
\alias{K1}
\title{1st cequilibrium constant of dissociation of carbonic acid in seawater}
\usage{
K1(TKelv, Sal)
}
\arguments{
\item{TKelv}{The temperature in Kelvin. You can use the function TCel() to calculate it}

\item{Sal}{The salinity in UPS}
}
\value{
The 1st equilibrium constant of dissociation
}
\description{
Calculates the 1st constant of dissociation of carbonic acid in seawater based on Zeebe and Wolf-Gladrow 2001, Mehrback et al. 1973 and Dickson and Millero 1987
}
