% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{K2}
\alias{K2}
\title{2nd equilibrium constant of dissociation of carbonic acid in seawater}
\usage{
K2(TKelv, Sal)
}
\arguments{
\item{TKelv}{The temperature in Kelvin. You can use the function TCel() to calculate it}

\item{Sal}{The salinity in UPS}
}
\value{
The 2nd equilibrium constant of dissociation
}
\description{
Calculated the 2nd constant of dissociation of carbonic acid in seawater based on Zeebe and Wolf-Gladrow 2001, Mehrback et al. 1973 and Dickson and Millero 1987
}
