% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{Kb}
\alias{Kb}
\title{Constant of base dissociation}
\usage{
Kb(Sal, TCel)
}
\arguments{
\item{Sal}{The salinity in UPS}

\item{TCel}{The temperature in degrees Celsius}
}
\value{
The constant of base dissociation
}
\description{
Calculates the constant of base dissociation based on Dickson et al. 1999, Zeebe and Wolf-Gladrow 2001
}
