% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{Kd}
\alias{Kd}
\title{Partition coefficient of B/Ca}
\usage{
Kd(protons)
}
\value{
The partition coefficient of B/Ca
}
\description{
Calculates the partition coefficient of B/Ca in mol/kg (McCulloch et al. 2017)
}
