% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{Ksp_arag}
\alias{Ksp_arag}
\title{Aragonite solubility constant}
\usage{
Ksp_arag(TKelv, Sal)
}
\arguments{
\item{TKelv}{The temperature in Kelvin. You can use the function TCel() to calculate it}

\item{Sal}{The salinity in UPS}
}
\value{
The aragonite solubility constant
}
\description{
Calculates the aragonite solubility constant in mol/kg based Zeebe and Wolf-Gladrow 2001
}
