% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{pHsw_cladocora}
\alias{pHsw_cladocora}
\title{Seawater pH reconstruction}
\usage{
pHsw_cladocora(pHcf)
}
\arguments{
\item{pHcf}{The pH of the calcifying fluid. You can use the function pHcf() to calculate it.}
}
\value{
The pHsw
}
\description{
Calculates the reconstructed pH of seawater using the species specific equation for the coral Cladocora caespitosa, defined by Trotter et al. 2011.
}
