% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{protons}
\alias{protons}
\title{Hydrogen ions concentrations}
\usage{
protons(pHcf)
}
\arguments{
\item{pHcf}{The pHcf. You can use the function pHcf() to calculate it}
}
\value{
The Hydrogen ions concentration
}
\description{
Calculates the concentration of hydrogen ions in mol/kg
}
