% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{Btot}
\alias{Btot}
\title{Total boron concentration}
\usage{
Btot(Sal)
}
\arguments{
\item{Sal}{The salinity in UPS}
}
\value{
The total boron concentration
}
\description{
Calculates the total boron concentration in mmol/kg based on Zeebe and Wolf-Gladrow 2001
}
\examples{
Btot1 <- Btot(35);
Btot2 <- Btot(c(35, 38));
}
