% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{carbonate_cf}
\alias{carbonate_cf}
\title{Carbonate ion concentration in the calcifying fluid}
\usage{
carbonate_cf(Kd, borate, BCa)
}
\arguments{
\item{Kd}{The partition coefficient of B/Ca. You can calculate it with the function carbonate()}

\item{borate}{The borate ion concentration. You can calculate it with the function borate()}

\item{BCa}{The B/Ca availability in the coral skeleton in mmol/mol}
}
\value{
The carbonate ion concentration
}
\description{
Calculates the concentration of carbonate ions in the calcifying fluid in umol/kg (based on McCulloch et al. 2017)
}
\examples{
CO32_cf <- carbonate_cf(2.97, 208.3, 664.8);
}
