% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{DICcf}
\alias{DICcf}
\title{Dissolved inorganic carbon of the calcifying fluid}
\usage{
DICcf(carbonate, protons, K1, K2)
}
\arguments{
\item{carbonate}{The concentration of carbonate ions in umol/kg. You can use the function carbonate() to calculate it.}

\item{protons}{The concentration of hydrogen ions in mol/kg. You can use the function protons() to calculate it.}

\item{K1}{The first equilibrium constant of dissociation of carbonic acid in seawater. You can use the function K1() to calculate it.}

\item{K2}{The second equilibrium constant of dissociation of carbonic acid in seawater. You can use the function K2() to calculate it.}
}
\value{
The DICcf
}
\description{
Calculates the DIC of the calcifying fluid based on D'Olivo and McCulloch 2017.
}
\examples{
DICcf_1 <- DICcf(930.5, 3.04e-09, 1.47e-06, 1.16e-09)
}
