% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{Kb_sw}
\alias{Kb_sw}
\title{Constant of base dissociation of boron in seawater}
\usage{
Kb_sw(Sal, TCel)
}
\arguments{
\item{Sal}{The salinity in UPS}

\item{TCel}{The temperature in degrees Celsius}
}
\value{
The constant of base dissociation
}
\description{
Calculates the constant of base dissociation of boron in seawater based on Dickson et al. 1999, Zeebe and Wolf-Gladrow 2001
}
