% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{Ksp_arag}
\alias{Ksp_arag}
\title{Aragonite solubility constant}
\usage{
Ksp_arag(TCel, Sal)
}
\arguments{
\item{TCel}{The temperature in degrees Celsius.}

\item{Sal}{The salinity in UPS.}
}
\value{
The aragonite solubility constant.
}
\description{
Calculates the aragonite solubility constant in mol/kg based Zeebe and Wolf-Gladrow 2001.
}
\examples{
Ksp1 <- Ksp_arag(24, 38)
Ksp2 <- Ksp_arag(c(23, 25, 27), c(37, 37.5, 38))
}
