% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{pHsw_coral}
\alias{pHsw_coral}
\title{Seawater pH reconstruction}
\usage{
pHsw_coral(pHcf, species = "Cladocora caespitosa")
}
\arguments{
\item{pHcf}{The pH of the calcifying fluid. You can use the function pHcf() to calculate it.}

\item{species}{Choose a coral species from the list: 
"Porites cylindrica" (Hönisch et al. 2004), 
"Acropora nobilis" (Hönisch et al. 2004),
"Acropora sp." (Reynaud et al. 2004),
"Porites sp." (Krief et al. 2010),
"Stylophora pistillata" (Krief et al. 2010),
"Cladocora caespitosa" (Trotter al. 2011)}
}
\value{
The pHsw
}
\description{
Calculates the reconstructed pH of seawater using the species specific equation for the coral Cladocora caespitosa, defined by Trotter et al. 2011.
}
\examples{
pHsw_Cladocora <- pHsw_coral(8.51, species = "Cladocora caespitosa");
}
