% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{protons_cf}
\alias{protons_cf}
\title{Hydrogen ions concentrations in the calcifying fluid}
\usage{
protons_cf(pHcf)
}
\arguments{
\item{pHcf}{The pHcf. You can use the function pHcf() to calculate it.}
}
\value{
The Hydrogen ions concentration.
}
\description{
Calculates the concentration of hydrogen ions in calcifying fluid in mol/kg.
}
\examples{
H_1 <- protons_cf(8.51)
}
