% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carb_parameters.R
\name{TCel_to_Kelv}
\alias{TCel_to_Kelv}
\title{Temperature conversion to kelvin}
\usage{
TCel_to_Kelv(TCel)
}
\arguments{
\item{TCel}{The temperature in degrees Celsius}
}
\value{
The temperature in kelvin
}
\description{
Converts temperature in Celsius to kelvin
}
\examples{
temp_kelvin <- TCel_to_Kelv(0);
temp_kelvin <- TCel_to_Kelv(c(0, -273));
}
